pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["ku2h7 <kukuhsyaifudin7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
poetry-dotenv-plugin = "^0.2.0"
supabase = "^2.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
